+++ UserFiles/BLE/SM/security_manager.h
@@ -45,11 +45,11 @@ typedef struct
 /* External functions declaration (Interface functions)         */
 /****************************************************************/
 uint16_t Get_Number_Of_Resolving_Records( void );
-uint16_t Add_Record_To_Resolving_List( RESOLVING_RECORD* Record );
-uint16_t Remove_Record_From_Resolving_List( IDENTITY_ADDRESS* Peer_Identity_Address );
+uint8_t Add_Record_To_Resolving_List( RESOLVING_RECORD* Record );
+uint8_t Remove_Record_From_Resolving_List( IDENTITY_ADDRESS* Peer_Identity_Address );
 RESOLVING_RECORD* Get_Record_From_Peer_Identity( IDENTITY_ADDRESS* Peer_Identity_Address );
 RESOLVING_RECORD* Get_Record_From_Index( uint16_t Index );
-uint16_t Clear_Resolving_List( void );
+uint8_t Clear_Resolving_List( void );
 
 
 /****************************************************************/
